+++ include/holyc-types.h
@@ -1,0 +1,10 @@
+typedef void                   U0;
+typedef unsigned char          U8;
+typedef unsigned short         U16;
+typedef unsigned int           U32;
+typedef unsigned long long int U64;
+
+typedef char          I8;
+typedef short         I16;
+typedef int           I32;
+typedef long long int I64;
