--- lib/cpu24/proc/std.h
@@ -1,5 +1,0 @@
-// Define processor type maps to a byte value.
-
-#define PROC_TYPE_GC16     0xFF
-#define PROC_TYPE_GC16X    0x00
-#define PROC_TYPE_GC24     0x01
