--- lib/gc-types.h
@@ -1,12 +1,0 @@
-/*
-  This file defines GC16X keyword types to
-  use in the CPU implementation (lib/cpu24/cpu24.h).
-*/
-
-// GC16X types
-typedef unsigned char  gcbyte;
-typedef unsigned short gcword;
-
-// Non-GC16X types (over 16 bits)
-typedef unsigned int           nongcdword;
-typedef unsigned long long int nongcqword;
